Moved Tensorflow-specific functions in tensorflow subpackage
