+++ tf_yarn/metrics.py
@@ -1,10 +1,5 @@
 import logging
-import skein
-import tensorflow as tf
-import os
-
 from datetime import timedelta
-
 from typing import (
     List,
     Optional,
@@ -13,7 +8,9 @@ from typing import (
     Tuple
 )
 
-from tf_yarn import mlflow, cluster
+import skein
+
+from tf_yarn import mlflow
 
 logger = logging.getLogger(__name__)
 
@@ -59,55 +56,3 @@ class OneShotMetricsLogger(NamedTuple):
             mlflow.set_tag(f"{mlflow.format_key(key)}_{self.n_try}", value)
             ret = True
         return ret
-
-
-class StepPerSecondHook(tf.estimator.StepCounterHook):
-
-    def __init__(
-        self,
-        every_n_steps=100,
-        every_n_secs=None,
-        output_dir=None,
-        summary_writer=None
-    ):
-        tf.estimator.StepCounterHook.__init__(
-            self,
-            every_n_steps=every_n_steps,
-            every_n_secs=every_n_secs,
-            output_dir=output_dir,
-            summary_writer=summary_writer
-        )
-
-    def _log_and_record(self, elapsed_steps: int, elapsed_time: float, global_step: int):
-        if cluster.is_chief():
-            steps_per_sec = elapsed_steps / elapsed_time
-            mlflow.log_metric(f"steps_per_sec_{cluster.n_try()}", steps_per_sec, step=global_step)
-
-
-def is_event_file(filename):
-    return os.path.basename(filename).startswith('events.out')
-
-
-def gen_events_iterator(model_path):
-    event_file = next((filename for filename in tf.compat.v1.gfile.ListDirectory(model_path)
-                       if is_event_file(filename)))
-    assert isinstance(event_file, str)
-    return tf.compat.v1.train.summary_iterator(os.path.join(model_path, event_file))
-
-
-def get_all_metrics(model_path):
-    events = gen_events_iterator(model_path)
-    dataframe = {
-        'step': list(),
-        'name': list(),
-        'value': list()
-    }
-    for event in events:
-        summary = event.summary
-        if summary:
-            for value in summary.value:
-                if value.simple_value:
-                    dataframe['step'].append(event.step)
-                    dataframe['name'].append(value.tag)
-                    dataframe['value'].append(value.simple_value)
-    return dataframe
